refactor(cli): dissolve clipboard action dispatcher
